requirements.txt
AnyQt==0.2.0
asgiref==3.8.1
cached-property==1.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
crispy-bootstrap5==2024.2
cycler==0.12.1
dj-database-url==2.2.0
Django==5.1
django-admin-volt==1.0.10
django-crispy-forms==2.3
django-filter==24.3
django-humanize==0.1.2
django-kronos==1.0
django-tastypie==0.14.7
django-url-filter==0.3.15
djangorestframework==3.15.2
enum-compat==0.0.3
Faker==28.1.0
fonttools==4.53.1
gunicorn==23.0.0
h11==0.14.0
humanize==4.10.0
idna==3.8
kiwisolver==1.4.5
kronos==0.2.2
Markdown==3.7
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyForms==3.0.0
PyOpenGL==3.1.7
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-mimeparse==1.6.0
QScintilla==2.14.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
visvis==1.14.0
whitenoise==6.7.0
